Upload header files to Git repository
This commit adds the header files to the Git repository. The following files have been included:

📦 Inc
 ┣ 📜 BIT_MATH.h
 ┣ 📜 CORE_NVIC_Config.h
 ┣ 📜 CORE_NVIC_Interface.h
 ┣ 📜 CORE_NVIC_Private.h
 ┣ 📜 MCAL_EXTI_Config.h
 ┣ 📜 MCAL_EXTI_Interface.h
 ┣ 📜 MCAL_EXTI_Private.h
 ┣ 📜 MCAL_GPIO_Config.h
 ┣ 📜 MCAL_GPIO_Interface.h
 ┣ 📜 MCAL_GPIO_Private.h
 ┣ 📜 MCAL_RCC_Config.h
 ┣ 📜 MCAL_RCC_Interface.h
 ┣ 📜 MCAL_RCC_Private.h
 ┣ 📜 MCAL_STK_Config.h
 ┣ 📜 MCAL_STK_Interface.h
 ┣ 📜 MCAL_STK_Private.h
 ┣ 📜 MCAL_TIM_Config.h
 ┣ 📜 MCAL_TIM_Interface.h
 ┣ 📜 MCAL_TIM_Private.h
 ┣ 📜 MCAL_USART_Config.h
 ┣ 📜 MCAL_USART_Interface.h
 ┣ 📜 MCAL_USART_Private.h
 ┣ 📜 STD_TYPES.h
 ┣ 📜 STM32F4xx_Device_Header.h
 ┗ 📜 STM32F4xx_ISR.h

These header files contain the necessary declarations, macros, and type definitions for the respective modules. They provide the interface and private definitions for NVIC, EXTI, GPIO, RCC, STK, TIM, USART, and other essential components of the project.

Including these header files in the repository ensures that the necessary declarations are available to other parts of the project during compilation and linking.

This commit complements the previously uploaded source files and sets the foundation for further development and collaboration.

Issue: #2