Add Doxygen style comments to MCAL_STK_Config.h.
Added Doxygen style comments to the top of the file, including a brief description of the file, the author, and the date.
Added Doxygen style comments to the #include statements, specifying the header files that are included.
Added Doxygen style comments to the STK_CLOCK_SOURCE macro, specifying its name, value, and possible values.