Added Doxygen comments This commit adds Doxygen style comments to the CORE_NVIC_Interface.h file. These comments improve the documentation for the file by making it more readable and understandable.
The following changes were made:

Doxygen's comments were added to the top of the file to provide a brief description of the file, the author, and the date.
Doxygen's comments were added to the #include statements to specify the header files that are included.
Doxygen's comments were added to the enumerations to specify the names of the enumerations and the possible values for each enumeration.
Doxygen's comments were added to the function declarations to specify the names of the functions, the parameters for each function, and the return value for each function.