Added Doxygen comments
Added Doxygen comments to the top of the file, including a brief description of the file, the author, and the date.
Added Doxygen comments to the #include statements, specifying the header files that are included.
Added Doxygen comments to the struct definitions, specifying the names of the structs and the members of each struct.
Added Doxygen comments to the #define macros, specifying the names of the macros and the values of each macro.

This commit improves the documentation for the CORE_NVIC_Private.h file by making it more readable and understandable.