Added Doxygen comments and omitted unnecessary comments.
Changelog
Added
Doxygen comments to the top of the file, including the author, date, and brief description of the file.
Doxygen comments to the #include statements, specifying the header files that are included.
Doxygen comments to the NUMOF_CONFIG_EXTI_SOURCES macro, specifying the name of the macro and the value of the macro.
Comments to the EXTI_GPIO_Mapping_t structure, explaining the purpose of each field.
Comments to the EXTI_PinCfg_t structure, explaining the purpose of each field.
Changed
Omitted the comments that are not necessary or that do not provide any additional information.
Added comments to explain the purpose of each function.
Removed
None.